feat: convert to multi-page Streamlit hub with generic landing page
- Rewrite app/app.py as generic AprovanLabs hub using st.navigation
- Extract fishing content into app/pages/fishing.py
- Remove fish-specific content from landing page
- Apply professional styling (reduce emojis, cleaner labels)
- Bump streamlit min version to >=1.36 for st.navigation
- Update smoke tests for multi-page architecture
