+++ pyproject.toml
@@ -14,7 +14,7 @@ matplotlib = ">=3.4.3,<4"
 pydantic = ">=2.0"
 requests = ">=2.28"
 great-expectations = "^0.18.0"
-streamlit = ">=1.28"
+streamlit = ">=1.36"
 streamlit-folium = ">=0.15"
 marimo = ">=0.11.0"
 folium = ">=0.14.0"
