+++ requirements.txt
@@ -1,4 +1,4 @@
-streamlit>=1.28
+streamlit>=1.36
 streamlit-folium>=0.15
 folium>=0.14
 pandas>1.5.0,<3.0.0
