+++ tests/test_app_smoke.py
@@ -17,11 +17,9 @@ import pytest
 REPO_ROOT = Path(__file__).resolve().parent.parent
 APP_DIR = REPO_ROOT / "app"
 COMPONENTS_DIR = APP_DIR / "components"
+PAGES_DIR = APP_DIR / "pages"
+FISHING_PAGE = PAGES_DIR / "fishing.py"
 
-
-# ---------------------------------------------------------------------------
-# 1. No `from app.*` imports anywhere in app/ (causes DuplicateElementId)
-# ---------------------------------------------------------------------------
 
 def _python_files_under(directory: Path) -> List[Path]:
     return sorted(p for p in directory.rglob("*.py") if p.is_file())
@@ -50,10 +48,6 @@ def test_no_app_prefix_import(filepath: Path) -> None:
             "Add app/ to sys.path and import from `components.*` instead."
         )
 
-
-# ---------------------------------------------------------------------------
-# 2. All app modules can be imported without Streamlit running
-# ---------------------------------------------------------------------------
 
 def test_dnr_map_importable() -> None:
     sys.path.insert(0, str(APP_DIR))
@@ -112,14 +106,12 @@ def test_onkia_models_importable() -> None:
     assert WaterTempPreference is not None
 
 
-# ---------------------------------------------------------------------------
-# 3. app.py is a single-page app (no pages/ directory)
-# ---------------------------------------------------------------------------
+def test_pages_directory_exists() -> None:
+    assert PAGES_DIR.is_dir(), "app/pages/ must exist for multi-page architecture"
 
-def test_no_pages_directory() -> None:
-    assert not (APP_DIR / "pages").is_dir(), (
-        "app/pages/ should not exist — the app is now a single page"
-    )
+
+def test_fishing_page_exists() -> None:
+    assert FISHING_PAGE.is_file(), "app/pages/fishing.py must exist"
 
 
 def test_app_py_parseable() -> None:
@@ -127,39 +119,49 @@ def test_app_py_parseable() -> None:
     ast.parse(source)
 
 
-# ---------------------------------------------------------------------------
-# 4. app.py has time-of-day selector
-# ---------------------------------------------------------------------------
+def test_fishing_page_parseable() -> None:
+    source = FISHING_PAGE.read_text()
+    ast.parse(source)
 
-def test_app_has_time_of_day_selector() -> None:
+
+def test_app_uses_st_navigation() -> None:
     source = (APP_DIR / "app.py").read_text()
-    assert "Time of Day" in source, "app.py must include a time-of-day selector"
+    assert "st.navigation" in source or "st.navigation" in source, (
+        "app.py must use st.navigation for multi-page routing"
+    )
 
 
-# ---------------------------------------------------------------------------
-# 5. app.py has depth profile and historical charts
-# ---------------------------------------------------------------------------
-
-def test_app_has_depth_profile() -> None:
+def test_app_no_fish_content() -> None:
     source = (APP_DIR / "app.py").read_text()
+    fish_terms = ["Wright County Fishing", "TARGET_SPECIES", "SPECIES_CODE_MAP"]
+    for term in fish_terms:
+        assert term not in source, (
+            f"app.py must not contain fish-specific content like '{term}' "
+            "-- it belongs in pages/fishing.py"
+        )
+
+
+def test_fishing_page_has_time_of_day_selector() -> None:
+    source = FISHING_PAGE.read_text()
+    assert "Time of Day" in source, "fishing.py must include a time-of-day selector"
+
+
+def test_fishing_page_has_depth_profile() -> None:
+    source = FISHING_PAGE.read_text()
     assert "depth_profile" in source.lower() or "depth" in source.lower()
 
 
-def test_app_has_usgs_glm_integration() -> None:
-    source = (APP_DIR / "app.py").read_text()
+def test_fishing_page_has_usgs_glm_integration() -> None:
+    source = FISHING_PAGE.read_text()
     assert "usgs_glm" in source.lower() or "get_temperature_profile" in source
     assert "thermocline" in source.lower()
     assert "species_depth_zones" in source.lower() or "depth_temp" in source.lower()
 
 
-def test_app_has_historical_temp_chart() -> None:
-    source = (APP_DIR / "app.py").read_text()
+def test_fishing_page_has_historical_temp_chart() -> None:
+    source = FISHING_PAGE.read_text()
     assert "cloud_cover" in source.lower() or "cloud" in source.lower()
 
-
-# ---------------------------------------------------------------------------
-# 6. requirements.txt exists and lists core deps
-# ---------------------------------------------------------------------------
 
 def test_requirements_txt_exists() -> None:
     assert (REPO_ROOT / "requirements.txt").is_file()
@@ -171,9 +173,12 @@ def test_requirements_includes_streamlit() -> None:
         assert dep in content.lower(), f"requirements.txt missing: {dep}"
 
 
-# ---------------------------------------------------------------------------
-# 7. No duplicate Streamlit element keys in app.py
-# ---------------------------------------------------------------------------
+def test_requirements_streamlit_min_version() -> None:
+    content = (REPO_ROOT / "requirements.txt").read_text()
+    assert "streamlit>=1.36" in content or "streamlit>=1.3" in content, (
+        "requirements.txt must require streamlit>=1.36 for st.navigation"
+    )
+
 
 def test_app_sidebar_keys_explicit() -> None:
     source = (APP_DIR / "app.py").read_text()
@@ -183,9 +188,13 @@ def test_app_sidebar_keys_explicit() -> None:
     )
 
 
-# ---------------------------------------------------------------------------
-# 8. DNR client has error handling
-# ---------------------------------------------------------------------------
+def test_fishing_page_sidebar_keys_explicit() -> None:
+    source = FISHING_PAGE.read_text()
+    button_calls = re.findall(r"st\.button\([^)]*key\s*=\s*['\"]([^'\"]+)['\"]", source)
+    assert len(button_calls) == len(set(button_calls)), (
+        f"Duplicate button keys in fishing.py: {button_calls}"
+    )
+
 
 def test_dnr_client_handles_connection_error() -> None:
     source = (REPO_ROOT / "src" / "onkia" / "dnr_client.py").read_text()
